refactor: use primitives TextField
